backend/requirements-phase1.txt
# Frontend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
aiosqlite==3.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
pyjwt==2.8.1
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# Rate limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0

# Async HTTP
aiohttp==3.9.1

# LLM Integration
ollama==0.0.11

# Utilities
python-dotenv==1.0.0

# Optional: CORS handling (already in FastAPI)
# Optional: WebSocket support for real-time features
websockets==12.0
